+++ packages/xrpl/snippets/src/bridgeTransfer.ts
@@ -20,11 +20,9 @@ async function sleep(sec: number): Promise<void> {
   })
 }
 
-const lockingClient = new Client(
-  'wss://sidechain-net1.devnet.rippletest.net:51233',
-)
+const lockingClient = new Client('wss://s.devnet.rippletest.net:51233')
 const issuingClient = new Client(
-  'wss://sidechain-net2.devnet.rippletest.net:51233',
+  'wss://devnetsidechain2.devnet.rippletest.net:51233',
 )
 const MAX_LEDGERS_WAITED = 5
 const LEDGER_CLOSE_TIME = 4
@@ -34,7 +32,7 @@ void bridgeTransfer()
 async function bridgeTransfer(): Promise<void> {
   await lockingClient.connect()
   await issuingClient.connect()
-  const lockingChainDoor = 'rMAXACCrp3Y8PpswXcg3bKggHX76V3F8M4'
+  const lockingChainDoor = 'rNQQyL2bJqbtgP5zXHJyQXamtrKYpgsbzV'
 
   const accountObjectsRequest: AccountObjectsRequest = {
     command: 'account_objects',
@@ -72,6 +70,7 @@ async function bridgeTransfer(): Promise<void> {
       parseInt(bridgeData.MinAccountCreateAmount as string, 10) * 2
     ).toString(),
   }
+  console.log(fundTx)
   const fundResponse = await lockingClient.submitAndWait(fundTx, {
     wallet: wallet1,
   })